requirements.txt
# Core LLM dependencies
openai>=1.0.0
anthropic>=0.18.0
mcp>=0.1.0

# Video processing
opencv-python>=4.8.0
ffmpeg-python>=0.2.0
moviepy>=1.0.3

# Image processing
Pillow>=10.0.0
torch>=2.0.0
transformers>=4.30.0

# API clients
requests>=2.31.0
aiohttp>=3.8.0

# Data management
sqlalchemy>=2.0.0
pydantic>=2.0.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0
